Add `add-metadata-hints` goal to Maven Plugin
Update the maven plugin with a new `add-metadata-hints` goal that can
be used to bundle hints obtained from the metadata repository into
the jar.
