+++ native-maven-plugin/src/main/java/org/graalvm/buildtools/maven/AbstractNativeMojo.java
@@ -41,416 +41,61 @@
 
 package org.graalvm.buildtools.maven;
 
-import org.apache.maven.artifact.Artifact;
-import org.apache.maven.execution.MavenSession;
-import org.apache.maven.plugin.AbstractMojo;
-import org.apache.maven.plugin.MojoExecution;
-import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.plugin.descriptor.PluginDescriptor;
-import org.apache.maven.plugins.annotations.Component;
-import org.apache.maven.plugins.annotations.Parameter;
-import org.apache.maven.project.MavenProject;
-import org.apache.maven.toolchain.ToolchainManager;
-import org.codehaus.plexus.logging.Logger;
-import org.graalvm.buildtools.Utils;
-import org.graalvm.buildtools.maven.config.ExcludeConfigConfiguration;
-import org.graalvm.buildtools.maven.config.MetadataRepositoryConfiguration;
-import org.graalvm.buildtools.utils.FileUtils;
-import org.graalvm.buildtools.utils.NativeImageUtils;
-import org.graalvm.buildtools.utils.SharedConstants;
-import org.graalvm.reachability.DirectoryConfiguration;
-import org.graalvm.reachability.GraalVMReachabilityMetadataRepository;
-import org.graalvm.reachability.internal.FileSystemRepository;
+import static org.graalvm.buildtools.utils.SharedConstants.METADATA_REPO_URL_TEMPLATE;
 
-import javax.inject.Inject;
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
-import java.nio.file.FileSystem;
-import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.function.Consumer;
 import java.util.function.Supplier;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-import static org.graalvm.buildtools.utils.SharedConstants.METADATA_REPO_URL_TEMPLATE;
+import javax.inject.Inject;
+
+import org.apache.maven.artifact.Artifact;
+import org.apache.maven.plugin.AbstractMojo;
+import org.apache.maven.plugins.annotations.Component;
+import org.apache.maven.plugins.annotations.Parameter;
+import org.apache.maven.project.MavenProject;
+import org.codehaus.plexus.logging.Logger;
+import org.graalvm.buildtools.maven.config.MetadataRepositoryConfiguration;
+import org.graalvm.buildtools.utils.FileUtils;
+import org.graalvm.buildtools.utils.SharedConstants;
+import org.graalvm.reachability.DirectoryConfiguration;
+import org.graalvm.reachability.GraalVMReachabilityMetadataRepository;
+import org.graalvm.reachability.internal.FileSystemRepository;
 
 /**
  * @author Sebastien Deleuze
  */
-
 public abstract class AbstractNativeMojo extends AbstractMojo {
-    protected static final String NATIVE_IMAGE_META_INF = "META-INF/native-image";
-    protected static final String NATIVE_IMAGE_PROPERTIES_FILENAME = "native-image.properties";
-    protected static final String NATIVE_IMAGE_DRY_RUN = "nativeDryRun";
-
-    @Parameter(defaultValue = "${plugin}", readonly = true) // Maven 3 only
-    protected PluginDescriptor plugin;
-
-    @Parameter(defaultValue = "${session}", readonly = true)
-    protected MavenSession session;
 
     @Parameter(defaultValue = "${project}", readonly = true, required = true)
     protected MavenProject project;
 
-    @Parameter(defaultValue = "${mojoExecution}")
-    protected MojoExecution mojoExecution;
-
-    @Parameter(property = "plugin.artifacts", required = true, readonly = true)
-    protected List<Artifact> pluginArtifacts;
-
-    @Parameter(defaultValue = "${project.build.directory}", property = "outputDir", required = true)
-    protected File outputDirectory;
-
-    @Parameter(property = "mainClass")
-    protected String mainClass;
-
-    @Parameter(property = "imageName", defaultValue = "${project.artifactId}")
-    protected String imageName;
-
-    @Parameter(property = "classpath")
-    protected List<String> classpath;
-
-    @Parameter(property = "classesDirectory")
-    protected File classesDirectory;
-
-    @Parameter(defaultValue = "${project.build.outputDirectory}", readonly = true, required = true)
-    protected File defaultClassesDirectory;
-
-    protected final List<Path> imageClasspath;
-
-    protected final Set<DirectoryConfiguration> metadataRepositoryConfigurations;
-
-    @Parameter(property = "debug", defaultValue = "false")
-    protected boolean debug;
-
-    @Parameter(property = "fallback", defaultValue = "false")
-    protected boolean fallback;
-
-    @Parameter(property = "verbose", defaultValue = "false")
-    protected boolean verbose;
-
-    @Parameter(property = "sharedLibrary", defaultValue = "false")
-    protected boolean sharedLibrary;
-
-    @Parameter(property = "quickBuild", defaultValue = "false")
-    protected boolean quickBuild;
-
-    @Parameter(property = "useArgFile")
-    protected Boolean useArgFile;
-
-    @Parameter(property = "buildArgs")
-    protected List<String> buildArgs;
-
-    @Parameter(defaultValue = "${project.build.directory}/native/generated", property = "resourcesConfigDirectory", required = true)
-    protected File resourcesConfigDirectory;
-
-    @Parameter(property = "agentResourceDirectory")
-    protected File agentResourceDirectory;
-
-    @Parameter(property = "excludeConfig")
-    protected List<ExcludeConfigConfiguration> excludeConfig;
-
-    @Parameter(property = "environmentVariables")
-    protected Map<String, String> environment;
-
-    @Parameter(property = "systemPropertyVariables")
-    protected Map<String, String> systemProperties;
-
-    @Parameter(property = "configurationFileDirectories")
-    protected List<String> configFiles;
-
-    @Parameter(property = "jvmArgs")
-    protected List<String> jvmArgs;
+    @Parameter(defaultValue = "${project.build.directory}/graalvm-reachability-metadata", required = true)
+    protected File reachabilityMetadataOutputDirectory;
 
     @Parameter(alias = "metadataRepository")
     protected MetadataRepositoryConfiguration metadataRepositoryConfiguration;
 
-    @Parameter(property = NATIVE_IMAGE_DRY_RUN, defaultValue = "false")
-    protected boolean dryRun;
+    protected final Set<DirectoryConfiguration> metadataRepositoryConfigurations;
 
     protected GraalVMReachabilityMetadataRepository metadataRepository;
 
     @Component
     protected Logger logger;
 
-    @Component
-    protected ToolchainManager toolchainManager;
-
     @Inject
     protected AbstractNativeMojo() {
-        imageClasspath = new ArrayList<>();
         metadataRepositoryConfigurations = new HashSet<>();
-        useArgFile = SharedConstants.IS_WINDOWS;
-    }
-
-    protected List<String> getBuildArgs() throws MojoExecutionException {
-        final List<String> cliArgs = new ArrayList<>();
-
-        if (excludeConfig != null) {
-            excludeConfig.forEach(entry -> {
-                cliArgs.add("--exclude-config");
-                cliArgs.add(entry.getJarPath());
-                cliArgs.add(String.format("\"%s\"", entry.getResourcePattern()));
-            });
-        }
-
-        cliArgs.add("-cp");
-        cliArgs.add(getClasspath());
-
-        if (debug) {
-            cliArgs.add("-g");
-        }
-        if (!fallback) {
-            cliArgs.add("--no-fallback");
-        }
-        if (verbose) {
-            cliArgs.add("--verbose");
-        }
-        if (sharedLibrary) {
-            cliArgs.add("--shared");
-        }
-
-        // Let's allow user to specify environment option to toggle quick build.
-        String quickBuildEnv = System.getenv("GRAALVM_QUICK_BUILD");
-        if (quickBuildEnv != null) {
-            logger.warn("Quick build environment variable is set.");
-            quickBuild = quickBuildEnv.isEmpty() || Boolean.parseBoolean(quickBuildEnv);
-        }
-
-        if (quickBuild) {
-            cliArgs.add("-Ob");
-        }
-
-        cliArgs.add("-H:Path=" + outputDirectory.toPath().toAbsolutePath());
-        cliArgs.add("-H:Name=" + imageName);
-
-        if (systemProperties != null) {
-            for (Map.Entry<String, String> entry : systemProperties.entrySet()) {
-                cliArgs.add("-D" + entry.getKey() + "=" + entry.getValue());
-            }
-        }
-
-        if (jvmArgs != null) {
-            jvmArgs.forEach(jvmArg -> cliArgs.add("-J" + jvmArg));
-        }
-
-        maybeAddGeneratedResourcesConfig(buildArgs);
-        maybeAddReachabilityMetadata(configFiles);
-
-        if (configFiles != null && !configFiles.isEmpty()) {
-            cliArgs.add("-H:ConfigurationFileDirectories=" +
-                    configFiles.stream()
-                    .map(Paths::get)
-                    .map(Path::toAbsolutePath)
-                    .map(Path::toString)
-                    .collect(Collectors.joining(","))
-            );
-        }
-
-        if (mainClass != null && !mainClass.equals(".")) {
-            cliArgs.add("-H:Class=" + mainClass);
-        }
-
-        if (buildArgs != null && !buildArgs.isEmpty()) {
-            for (String buildArg : buildArgs) {
-                cliArgs.addAll(Arrays.asList(buildArg.split("\\s+")));
-            }
-        }
-
-        if (useArgFile) {
-            Path tmpDir = Paths.get("target", "tmp");
-            return NativeImageUtils.convertToArgsFile(cliArgs, tmpDir);
-        }
-        return Collections.unmodifiableList(cliArgs);
-    }
-
-    protected Path processSupportedArtifacts(Artifact artifact) throws MojoExecutionException {
-        return processArtifact(artifact, "jar", "test-jar", "war");
-    }
-
-    protected Path processArtifact(Artifact artifact, String... artifactTypes) throws MojoExecutionException {
-        File artifactFile = artifact.getFile();
-
-        if (artifactFile == null) {
-            logger.debug("Missing artifact file for artifact " + artifact + " (type: " + artifact.getType() + ")");
-            return null;
-        }
-
-        if (Arrays.stream(artifactTypes).noneMatch(a -> a.equals(artifact.getType()))) {
-            logger.warn("Ignoring ImageClasspath Entry '" + artifact + "' with unsupported type '" + artifact.getType() + "'");
-            return null;
-        }
-        if (!artifactFile.exists()) {
-            throw new MojoExecutionException("Missing jar-file for " + artifact + ". " +
-                    "Ensure that " + plugin.getArtifactId() + " runs in package phase.");
-        }
-
-        Path jarFilePath = artifactFile.toPath();
-        logger.debug("ImageClasspath Entry: " + artifact + " (" + jarFilePath.toUri() + ")");
-
-        warnIfWrongMetaInfLayout(jarFilePath, artifact);
-        return jarFilePath;
-    }
-
-    protected void addArtifactToClasspath(Artifact artifact) throws MojoExecutionException {
-        Optional.ofNullable(processSupportedArtifacts(artifact)).ifPresent(imageClasspath::add);
-    }
-
-    protected void warnIfWrongMetaInfLayout(Path jarFilePath, Artifact artifact) throws MojoExecutionException {
-        if (jarFilePath.toFile().isDirectory()) {
-            logger.debug("Artifact `" + jarFilePath + "` is a directory.");
-            return;
-        }
-        URI jarFileURI = URI.create("jar:" + jarFilePath.toUri());
-        try (FileSystem jarFS = FileSystems.newFileSystem(jarFileURI, Collections.emptyMap())) {
-            Path nativeImageMetaInfBase = jarFS.getPath("/" + NATIVE_IMAGE_META_INF);
-            if (Files.isDirectory(nativeImageMetaInfBase)) {
-                try (Stream<Path> stream = Files.walk(nativeImageMetaInfBase)) {
-                    List<Path> nativeImageProperties = stream
-                            .filter(p -> p.endsWith(NATIVE_IMAGE_PROPERTIES_FILENAME)).collect(Collectors.toList());
-                    for (Path nativeImageProperty : nativeImageProperties) {
-                        Path relativeSubDir = nativeImageMetaInfBase.relativize(nativeImageProperty).getParent();
-                        boolean valid = relativeSubDir != null && (relativeSubDir.getNameCount() == 2);
-                        valid = valid && relativeSubDir.getName(0).toString().equals(artifact.getGroupId());
-                        valid = valid && relativeSubDir.getName(1).toString().equals(artifact.getArtifactId());
-                        if (!valid) {
-                            String example = NATIVE_IMAGE_META_INF + "/%s/%s/" + NATIVE_IMAGE_PROPERTIES_FILENAME;
-                            example = String.format(example, artifact.getGroupId(), artifact.getArtifactId());
-                            logger.warn("Properties file at '" + nativeImageProperty.toUri() + "' does not match the recommended '" + example + "' layout.");
-                        }
-                    }
-                }
-            }
-        } catch (IOException e) {
-            throw new MojoExecutionException("Artifact " + artifact + "cannot be added to image classpath", e);
-        }
-    }
-
-    protected abstract List<String> getDependencyScopes();
-
-    protected void addDependenciesToClasspath() throws MojoExecutionException {
-        configureMetadataRepository();
-        for (Artifact dependency : project.getArtifacts().stream()
-                .filter(artifact -> getDependencyScopes().contains(artifact.getScope()))
-                .collect(Collectors.toSet())) {
-            addArtifactToClasspath(dependency);
-            maybeAddDependencyMetadata(dependency);
-        }
-    }
-
-    /**
-     * Returns path to where application classes are stored, or jar artifact if it is produced.
-     * @return Path to application classes
-     * @throws MojoExecutionException failed getting main build path
-     */
-    protected Path getMainBuildPath() throws MojoExecutionException {
-        if (classesDirectory != null) {
-            return classesDirectory.toPath();
-        } else {
-            Path artifactPath = processArtifact(project.getArtifact(), project.getPackaging());
-            if (artifactPath != null) {
-                return artifactPath;
-            } else {
-                return defaultClassesDirectory.toPath();
-            }
-        }
-    }
-
-    protected void populateApplicationClasspath() throws MojoExecutionException {
-        imageClasspath.add(getMainBuildPath());
-    }
-
-    protected void populateClasspath() throws MojoExecutionException {
-        if (classpath != null && !classpath.isEmpty()) {
-            imageClasspath.addAll(classpath.stream()
-                    .map(Paths::get)
-                    .map(Path::toAbsolutePath)
-                    .collect(Collectors.toSet())
-            );
-        } else {
-            populateApplicationClasspath();
-            addDependenciesToClasspath();
-        }
-        imageClasspath.removeIf(entry -> !entry.toFile().exists());
-    }
-
-    protected String getClasspath() throws MojoExecutionException {
-        populateClasspath();
-        if (imageClasspath.isEmpty()) {
-            throw new MojoExecutionException("Image classpath is empty. " +
-                    "Check if your classpath configuration is correct.");
-        }
-        return imageClasspath.stream()
-                .map(Path::toString)
-                .collect(Collectors.joining(File.pathSeparator));
-    }
-
-    protected void buildImage() throws MojoExecutionException {
-        Path nativeImageExecutable = Utils.getNativeImage(logger);
-
-        try {
-            ProcessBuilder processBuilder = new ProcessBuilder(nativeImageExecutable.toString());
-            processBuilder.command().addAll(getBuildArgs());
-
-            if (environment != null) {
-                processBuilder.environment().putAll(environment);
-            }
-
-            if (!outputDirectory.exists() && !outputDirectory.mkdirs()) {
-                throw new MojoExecutionException("Failed creating output directory");
-            }
-            processBuilder.inheritIO();
-
-            String commandString = String.join(" ", processBuilder.command());
-            logger.info("Executing: " + commandString);
-
-            if (dryRun) {
-                logger.warn("Skipped native-image building due to `" + NATIVE_IMAGE_DRY_RUN + "` being specified.");
-                return;
-            }
-
-            Process imageBuildProcess = processBuilder.start();
-            if (imageBuildProcess.waitFor() != 0) {
-                throw new MojoExecutionException("Execution of " + commandString + " returned non-zero result");
-            }
-        } catch (IOException | InterruptedException e) {
-            throw new MojoExecutionException("Building image with " + nativeImageExecutable + " failed", e);
-        }
-    }
-
-    protected void maybeAddGeneratedResourcesConfig(List<String> into) {
-        if (resourcesConfigDirectory.exists() || agentResourceDirectory != null) {
-            File[] dirs = resourcesConfigDirectory.listFiles();
-            Stream<File> configDirs =
-                    Stream.concat(dirs == null ? Stream.empty() : Arrays.stream(dirs),
-                            agentResourceDirectory == null ? Stream.empty() : Stream.of(agentResourceDirectory).filter(File::isDirectory));
-
-            String value = configDirs.map(File::getAbsolutePath).collect(Collectors.joining(","));
-            if (!value.isEmpty()) {
-                into.add("-H:ConfigurationFileDirectories=" + value);
-                if (agentResourceDirectory != null && agentResourceDirectory.isDirectory()) {
-                    // The generated reflect config file contains references to java.*
-                    // and org.apache.maven.surefire that we'd need to remove using
-                    // a proper JSON parser/writer instead
-                    into.add("-H:+AllowIncompleteClasspath");
-                }
-            }
-        }
     }
 
     protected boolean isMetadataRepositoryEnabled() {
@@ -460,7 +105,7 @@ public abstract class AbstractNativeMojo extends AbstractMojo {
     protected void configureMetadataRepository() {
         if (isMetadataRepositoryEnabled()) {
             Path repoPath = null;
-            Path destinationRoot = outputDirectory.toPath().resolve("graalvm-reachability-metadata");
+            Path destinationRoot = reachabilityMetadataOutputDirectory.toPath();
             if (Files.exists(destinationRoot) && !Files.isDirectory(destinationRoot)) {
                 throw new RuntimeException("Metadata repository must be a directory, please remove regular file at: " + destinationRoot);
             }
@@ -472,7 +117,7 @@ public abstract class AbstractNativeMojo extends AbstractMojo {
 
             if (metadataRepositoryConfiguration.getLocalPath() != null) {
                 Path localPath = metadataRepositoryConfiguration.getLocalPath().toPath();
-                Path destination = outputDirectory.toPath().resolve(FileUtils.hashFor(localPath.toUri()));
+                Path destination = destinationRoot.resolve(FileUtils.hashFor(localPath.toUri()));
                 repoPath = unzipLocalMetadata(localPath, destination);
             } else {
                 URL targetUrl = metadataRepositoryConfiguration.getUrl();
@@ -527,17 +172,7 @@ public abstract class AbstractNativeMojo extends AbstractMojo {
         }
     }
 
-    protected void maybeAddReachabilityMetadata(List<String> configDirs) {
-        if (isMetadataRepositoryEnabled() && !metadataRepositoryConfigurations.isEmpty()) {
-            metadataRepositoryConfigurations.stream()
-                    .map(configuration -> configuration.getDirectory().toAbsolutePath())
-                    .map(Path::toFile)
-                    .map(File::getAbsolutePath)
-                    .forEach(configDirs::add);
-        }
-    }
-
-    protected void maybeAddDependencyMetadata(Artifact dependency) {
+    protected void maybeAddDependencyMetadata(Artifact dependency, Consumer<File> excludeAction) {
         if (isMetadataRepositoryEnabled() && metadataRepository != null && !isArtifactExcludedFromMetadataRepository(dependency)) {
             Set<DirectoryConfiguration> configurations = metadataRepository.findConfigurationsFor(q -> {
                 q.useLatestConfigWhenVersionIsUntested();
@@ -550,10 +185,8 @@ public abstract class AbstractNativeMojo extends AbstractMojo {
                 });
             });
             metadataRepositoryConfigurations.addAll(configurations);
-            if (configurations.stream().anyMatch(DirectoryConfiguration::isOverride)) {
-                buildArgs.add("--exclude-config");
-                buildArgs.add(Pattern.quote(dependency.getFile().getAbsolutePath()));
-                buildArgs.add("^/META-INF/native-image/");
+            if (excludeAction != null && configurations.stream().anyMatch(DirectoryConfiguration::isOverride)) {
+                excludeAction.accept(dependency.getFile());
             }
         }
     }
